+++ src/index.ts
@@ -292,6 +292,7 @@ export {
 export type {
   OnModuleInit,
   OnApplicationBootstrap,
+  OnFirstRequest,
   OnModuleDestroy,
   OnApplicationShutdown,
   BeforeApplicationShutdown,